fix: failed requests should always be rescheduled
Previously rescheduling only handled the case of a failed fetch
operation, ie. network or http related errors like being rate-limited.
As we always want to retry under these circumstances, preventing
previously retried tasks to be rescheduled again is wrong, hence the
change. Also, with this change there is no point in the
reschedule_counter anymore, at least as long there is no way for a task
to be rescheduled outside of fetching.